Add `glob` to `dir-dependency` messages
