+++ src/jit/processTailwindFeatures.js
@@ -8,26 +8,17 @@ import { createContext } from './lib/setupContextUtils'
 
 export default function processTailwindFeatures(setupContext) {
   return function (root, result) {
-    function registerDependency(dependency) {
-      // rollup-plugin-postcss does not support dir-dependency messages
-      // but directories can be watched in the same way as files
-      if (dependency.dir && process.env.ROLLUP_WATCH === 'true') {
-        dependency = { file: dependency.dir }
-      }
-
-      result.messages.push({
-        type: dependency.dir ? 'dir-dependency' : 'dependency',
-        plugin: 'tailwindcss',
-        parent: result.opts.from,
-        ...dependency,
-      })
-    }
-
     let tailwindDirectives = normalizeTailwindDirectives(root)
 
     let context = setupContext({
       tailwindDirectives,
-      registerDependency,
+      registerDependency(dependency) {
+        result.messages.push({
+          plugin: 'tailwindcss',
+          parent: result.opts.from,
+          ...dependency,
+        })
+      },
       createContext(tailwindConfig, changedContent) {
         return createContext(tailwindConfig, changedContent, tailwindDirectives, root)
       },
